requirements.txt
Pillow==9.4.0
razorpay==1.3.0
requests==2.28.1
streamlit==1.16.0
streamlit_lottie==0.0.3
